changes done

+++ Wedsimple Project/Event/EventViewController.m
@@ -118,7 +118,7 @@
         [cell.checkImgv setImage:[UIImage imageNamed:[checkImage objectAtIndex:indexPath.row]]];
     }
     
-    cell.EventLbl.text = [NSString stringWithFormat:@"%d",indexPath.row];
+    cell.EventLbl.text = [NSString stringWithFormat:@"%ld",(long)indexPath.row];
     
     
     [cell.editBtn addTarget:self action:@selector(Edit:) forControlEvents:UIControlEventTouchUpInside];
@@ -140,7 +140,7 @@
 }
 -(void) Edit:(UIButton*)button
 {
-    NSLog(@"%d",button.tag);
+    NSLog(@"%ld",(long)button.tag);
 }
 
 -(void)image
